feat: 자동차 전진 여부에 대한 시도 별 결과 출력 기능 추가
