+++ .eslintrc.json
@@ -1,6 +1,7 @@
 {
   "rules": {
-    "linebreak-style": 0
+    "linebreak-style": 0,
+    "max-lines-per-function": ["error", { "max" : 10 }]
   },
   "env": {
     "es6": true,
