+++ src/RacingCarGame/Car.js
@@ -24,6 +24,10 @@ class Car {
     }
   }
 
+  info() {
+    return { name: this.#name, position: this.#position };
+  }
+
   //test를 위한 임의 getter
   position() {
     return this.#position;
