+++ src/view/OutputView.js
@@ -1,6 +1,20 @@
+import MESSAGE from "../constants/Message.js";
+
 class OutputView {
   static printError(error) {
     console.log(error.message);
+  }
+
+  static printCar(car) {
+    console.log(MESSAGE.printCarInfo(car));
+  }
+
+  static printQuery(string) {
+    console.log(string);
+  }
+
+  static newLine() {
+    console.log("");
   }
 }
